requirements.txt
spyse.py
requests
colorama